requirements.txt
httpx==0.24.1
python-dotenv==1.0.0
vercel==31.0.1
pytest==7.3.1